Set FS.fetch to a test double to make sure we don't use real method. (#22)
* Set FS.fetch to a test double to make sure we don't use real method.

Without this method, there is a timing hole were elements get loaded
before the stub is set. It's important when testing that
fs-globals-mock.js loads before fs-globals.js so that it sets FS.fetch
first. The code checks whether it is already set by the mock and will
not override it, if it is.

* Only stub window.fetch if it is defined.

* Enhance check to include window and window.fetch.